requirements.txt
torch
torchvision
transformers
pytesseract
Pillow
pandas
scikit-learn
fastapi
uvicorn
python-multipart
requests
google-generativeai
python-dotenv
pytest
tqdm
matplotlib
opencv-python-headless
python-pptx
